backend/pyproject.toml
[project]
name = "BrainCard"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "diklios", email = "diklios5768@gmail.com"},
]
dependencies = [
    "flask>=3.0.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "gevent>=24.2.1",
    "gunicorn>=22.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
